URL Builder for Kobo/Commcare Integrations (#578)
* Do the thing

* Clean

* 🐜  it!

* bump
