+++ app/client/components/pages/settings/settingsPage.jsx
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom";
 
 import AuthModule from "../../auth/authModule.jsx";
 import UserList from "../../adminUser/adminUserList.jsx";
-import KoboCredentials from "../../koboCredentials";
+import ExternalAuthCredentials from "../../externalAuthCredentials";
 import GetVerified from "../../GetVerified";
 
 import {
@@ -24,7 +24,7 @@ export default class settingsPage extends React.Component {
             <RestrictedModuleBox>
               <ModuleHeader>SETTINGS</ModuleHeader>
               <p style={{ margin: "1em" }}></p>
-              <KoboCredentials />
+              <ExternalAuthCredentials />
               <GetVerified />
               <PageLink
                 to="/settings/organisation"
